requirements.txt
# =============================================================================
# AARA - ANIME DESKTOP ASSISTANT DEPENDENCIES
# =============================================================================
# Strategy: Premium services first → Free fallback
# Organized by component

# =============================================================================
# CORE (Always Required)
# =============================================================================
PySide6>=6.6.0
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.26.0
pillow>=10.0.0
requests>=2.31.0

# =============================================================================
# LLM PROVIDERS
# =============================================================================
# PREMIUM: Anthropic Claude
anthropic>=0.45.0

# FREE: Groq (cloud, free tier)
groq>=0.4.0

# FREE: Google Gemini
google-generativeai>=0.3.0

# For OpenRouter and OpenAI-compatible APIs
openai>=1.10.0

# =============================================================================
# TEXT-TO-SPEECH (TTS)
# =============================================================================
# PREMIUM: ElevenLabs
elevenlabs>=1.0.0

# FREE: Microsoft Edge TTS (recommended free option)
edge-tts>=6.1.0

# FALLBACK: Offline pyttsx3
pyttsx3>=2.90

# =============================================================================
# SPEECH-TO-TEXT (STT)
# =============================================================================
# Whisper (local) - best quality
openai-whisper>=20231117

# Vosk (lightweight, offline)
vosk>=0.3.45

# Audio capture
sounddevice>=0.4.6
soundfile>=0.12.1

# =============================================================================
# WAKE WORD / ACTIVATION
# =============================================================================
# PREMIUM: Picovoice Porcupine
pvporcupine>=3.0.0

# FREE: Keyboard hotkey detection
pynput>=1.7.6

# =============================================================================
# WEB SEARCH
# =============================================================================
# FREE: DuckDuckGo
duckduckgo-search>=4.0.0

# =============================================================================
# MEMORY & DATABASE
# =============================================================================
# Vector database for long-term memory
chromadb>=0.4.0

# SQLite is built-in to Python

# =============================================================================
# SYSTEM CONTROL
# =============================================================================
# Mouse/keyboard automation
pyautogui>=0.9.54

# Window management (Windows)
pygetwindow>=0.0.9

# Volume control (Windows)
pycaw>=20230407

# Clipboard
pyperclip>=1.8.2

# =============================================================================
# UTILITIES
# =============================================================================
colorama>=0.4.6
aiohttp>=3.9.0
aiofiles>=23.2.1
